prepare truncate test
